% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frederik_ATV.R
\name{Frederik_ATV}
\alias{Frederik_ATV}
\title{Frederik Add Two Vectors}
\usage{
Frederik_ATV(first_vector, second_vector)
}
\arguments{
\item{first_vector}{numeric vector of size n}

\item{second_vector}{numeric vector of size n}
}
\value{
the vector resulting from adding the two input vectors together. Numeric with size n
}
\description{
Adds two vectors together
}
\examples{
Frederik_ATV(c(1.5,2.5,3.5), c(4.5, 5.5, 6.5))
}
