% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frederik_hclust.R
\name{Frederik_hclust}
\alias{Frederik_hclust}
\title{Hierarchical Clustering Algorithm}
\usage{
Frederik_hclust(matrix_of_data, target_nr_of_clusters)
}
\arguments{
\item{matrix_of_data}{the input dataset. Must be a purely numeric matrix [observations x dimentions]}

\item{target_nr_of_clusters}{the desired number of clusters}
}
\value{
a vector of length [observations] showing clusterings for all datapoints in the input data matrix
}
\description{
Computes clusterings for a dataset via the hierarchical clustering method
}
\examples{
data_set = as.matrix(iris[, -5])
Frederik_hclust(data_set, 3)
}
